+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.1.7](https://github.com/bitrix24/b24ui/compare/v2.1.6...v2.1.7) (2025-11-xx)
+
+* docs(installation/vue): typo fix
+* docs(typography/CardGroup): typo fix
+* docs(ComponentCode): add missing cast imports
+
 ## [2.1.6](https://github.com/bitrix24/b24ui/compare/v2.1.5...v2.1.6) (2025-11-20)
 
 ### Bug Fixes